requirements.txt
fastapi
uvicorn
pydantic
chromadb
numpy<2.0
jinja2
python-dotenv
beautifulsoup4
requests
langchain
langchain-core
langchain-google-genai
langchain-huggingface
langchain-chroma
langgraph
sentence-transformers
